.claude/agents/testing-agent/requirements.txt
# Claude Agent SDK
claude-agent-sdk>=0.1.0

# Environment management
python-dotenv>=1.0.0

# HTTP requests (for validation/testing utilities)
requests>=2.31.0
aiohttp>=3.9.0

# Testing frameworks (optional, for unit testing the agent itself)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# JSON schema validation
jsonschema>=4.20.0

# Pretty printing and reporting
rich>=13.7.0
tabulate>=0.9.0
